Isolating sub-model classes in a separate file
